fetchData allow custom filter params.
Many other places where you need to load an table, some values can be added to the params of the table, so the BE can filter this data and reduce the number of data sent. also query param can be set as part of the options before the table loads and the be filters.